Replace shoddy FNV-1a hashing with xxHash3
- Use the streaming hash object for when we're using complex composite struct types that can't be cast to a good ol span of bytes
- Use thead_local streaming hash structs here, just in case we thread this in the future. Mostly means we aren't paying alloc cost each invocation though. For now.
- Have most functions go through a HashBytes that just takes structs as a span of bytes
- Refactor a few structs to be just that, spans of bytes aligned right to not have hidden padding (because then hashing would be UB)
- Widen a few enums for the same reason

+++ client/include/ShaderLibraryTypes.hpp
@@ -97,8 +97,10 @@ enum class SamplerBindingType : uint8_t
     Comparison,
 };
 
-/** @brief The scalar type of one vertex attribute or one color target. */
-enum class VertexScalarType : uint8_t
+/** @brief The scalar type of one vertex attribute or one color target.
+ *  @note Widened to uint32_t so struct holding it is 16 bytes even vs 13
+*/
+enum class VertexScalarType : uint32_t
 {
     Invalid = 0,
     Float16,
